+++ demos/camera/src/main/java/io/github/thibaultbee/streampack/app/data/storage/DataStoreRepository.kt
@@ -115,7 +115,7 @@ class DataStoreRepository(
             mimeType = mimeType,
             startBitrate = startBitrate,
             resolution = resolution,
-            fps = fps,
+            fps = fps.toFloat(),
             profile = profile,
             level = level
         )
